+++ modules/alloy/changelog.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 # [5.0.2] - 2019-11-11
 
 ### Fixed
+- Fixed `TouchDragging` behaviour triggering drag on any document touchmove event.
+- Fixed dragging updating start state on window scroll, when dragging hadn't started.
 - Fixed menu item execute not killing the original mouse or touch event.
 - Fixed touchstart events bubbling up from buttons.